temp plan improvements
